review algorithm, create separated module to check angle brackets, compartimentalized methods on validateFile class
